% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reddit.R
\name{get_user_comments}
\alias{get_user_comments}
\title{Get user content from reddit}
\usage{
get_user_comments(
  user_name,
  limit = 10L,
  after = NULL,
  base_url = "https://www.reddit.com/"
)
}
\arguments{
\item{user_name}{character string for an individual reddit user}

\item{limit}{integer value between 1 and 100}

\item{after}{character string specifying the location in a list of user comments. If specified, the comments after that location are returned (if found).}

\item{base_url}{base url for the GET request}
}
\value{
list containing named items: `data`, `after`, and `request`.
`data` is a data frame of reddit comments for a user.
`after` is a character value of the hashed location of the last comment included in the returned items. This value can be used as a starting point for the next batch of responses.
`request` is the exact string for which results are returned.
}
\description{
Download an individual user's content from reddit
}
\examples{
# example code
sprog2 <- get_user_comments(user_name="poem_for_your_sprog", limit=2)
sprog_next2 <- get_user_comments(user_name="poem_for_your_sprog", limit=2, after=sprog2$after)
sprog4 <- get_user_comments(user_name="poem_for_your_sprog", limit=4)
}
