% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earthquakes.R
\name{get_earthquakes}
\alias{get_earthquakes}
\title{Get earthquake information from USGS}
\usage{
get_earthquakes(
  start_time,
  end_time = NULL,
  min_magnitude = NULL,
  base_url = "https://earthquake.usgs.gov/fdsnws/event/1/query"
)
}
\arguments{
\item{start_time}{character string specifying the start date for the records in the form 'yyyy-mm-dd'}

\item{end_time}{character string specifying the end date for the records in the form 'yyyy-mm-dd'}

\item{min_magnitude}{numeric quantity specifying the minimum magnitude considered for a record}

\item{base_url}{base url for the GET request}
}
\value{
data frame of earth quake records. Detailed information on each of the variables can be found as part of the [USGS ComCat Documentation](https://earthquake.usgs.gov/data/comcat/index.php)
}
\description{
Downloads records of earth quakes from the USGS Earthquake Catalog
}
\examples{
# example code
one_day <- get_earthquakes(start_time = "2025-03-06", end_time = "2025-03-07")
# all documented earthquakes with a magnitude of at least 9 on the Richter scale
nines <-  get_earthquakes(start_time = "1800-01-01", end_time = "2025-03-07", min_magnitude=9)
}
